image similarity metrics with openCV
