+++ src/opencv/opencv.d.ts
@@ -1,0 +1,3 @@
+declare module 'opencv-4.8.0.js' {
+  export const cv: any;
+}
